dev-1.2.2.47
You can no longer use the chat (or use commands like /msg and similar) in the time when you're answering a trivia bot question.

+++ src/main/java/net/mat0u5/lifeseries/client/trivia/gui/QuizScreen.java
@@ -39,7 +39,7 @@ public class QuizScreen extends Screen {
 
     private final List<List<OrderedText>> answers = new ArrayList<>();
     private String DIFFICULTY = "Difficulty: null";
-    private int timerSeconds = 120;
+    private long timerSeconds = 120;
     private final List<Rectangle> answerRects = new ArrayList<>();
 
     public QuizScreen() {
@@ -176,8 +176,8 @@ public class QuizScreen extends Screen {
         */
 
         // Timer
-        int minutes = timerSeconds / 60;
-        int seconds = timerSeconds - minutes * 60;
+        long minutes = timerSeconds / 60;
+        long seconds = timerSeconds - minutes * 60;
         String secondsStr = String.valueOf(seconds);
         String minutesStr = String.valueOf(minutes);
         while (secondsStr.length() < 2) secondsStr = "0" + secondsStr;
